+++ ubelt/__init__.py
@@ -26,7 +26,7 @@ Testing:
     xdoctest ubelt
 """
 
-__version__ = '1.0.2'
+__version__ = '1.1.0'
 
 __submodules__ = {
     'util_arg': None,